Use sodium_malloc/sodium_free + RAII to clean-up memory from sensible data

+++ src/vault/vault/common/retcodes.h
@@ -1,9 +1,7 @@
 #ifndef ERRCODES_H
 #define ERRCODES_H
 
-#include <cstdint>
-
-using VaultReturnCode = uint8_t;
+using VaultReturnCode = int;
 
 constexpr VaultReturnCode VAULT_SUCCESS = 0;
 constexpr VaultReturnCode VAULT_GENERIC_ERROR = 1;
